MBP-447 - Project Roles with data access restrictions  * Finish off security for workflows

+++ biocode-server/src/com/biomatters/plugins/biocode/server/Workflows.java
@@ -1,7 +1,14 @@
 package com.biomatters.plugins.biocode.server;
 
 import com.biomatters.geneious.publicapi.databaseservice.DatabaseServiceException;
+import com.biomatters.geneious.publicapi.databaseservice.Query;
+import com.biomatters.geneious.publicapi.documents.Condition;
+import com.biomatters.plugins.biocode.labbench.BiocodeService;
+import com.biomatters.plugins.biocode.labbench.PlateDocument;
 import com.biomatters.plugins.biocode.labbench.Workflow;
+import com.biomatters.plugins.biocode.labbench.WorkflowDocument;
+import com.biomatters.plugins.biocode.labbench.lims.LIMSConnection;
+import com.biomatters.plugins.biocode.labbench.lims.LimsSearchResult;
 import com.biomatters.plugins.biocode.server.security.AccessUtilities;
 import com.biomatters.plugins.biocode.server.security.Role;
 import com.biomatters.plugins.biocode.server.utilities.RestUtilities;
@@ -9,6 +16,8 @@ import com.biomatters.plugins.biocode.server.utilities.RestUtilities;
 import javax.ws.rs.*;
 import javax.ws.rs.core.NoContentException;
 import java.util.Collections;
+import java.util.HashSet;
+import java.util.List;
 
 /**
  * @author Matthew Cheung
@@ -27,7 +36,7 @@ public class Workflows {
         }
         try {
             Workflow result = RestUtilities.getOnlyItemFromList(LIMSInitializationListener.getLimsConnection().getWorkflows(Collections.singletonList(id)), "No workflow for id: " + id);
-            AccessUtilities.checkUserHasRoleForSamples(Collections.singletonList(result.getFimsSample()), Role.READER);
+            AccessUtilities.checkUserHasRoleForExtractionIds(Collections.singletonList(result.getExtractionId()), Role.READER);
             return result;
         } catch (DatabaseServiceException e) {
             throw new InternalServerErrorException(e.getMessage(), e);
@@ -38,8 +47,8 @@ public class Workflows {
     @Consumes("text/plain")
     @Path("{id}/name")
     public void renameWorkflow(@PathParam("id") int id, String newName) {
-        // todo get workflow and check access
         try {
+            checkAccessForWorkflowId(id, Role.WRITER);
             LIMSInitializationListener.getLimsConnection().renameWorkflow(id, newName);
         } catch (DatabaseServiceException e) {
             throw new WebApplicationException(e.getMessage(), e);
@@ -49,11 +58,25 @@ public class Workflows {
     @DELETE
     @Path("{workflowId}/sequences/{extractionId}")
     public void deleteSequencesForWorkflow(@PathParam("workflowId")int workflowId, @PathParam("extractionId")String extractionId) {
-        // todo get workflow and check access
         try {
+            checkAccessForWorkflowId(workflowId, Role.WRITER);
             LIMSInitializationListener.getLimsConnection().deleteSequencesForWorkflowId(workflowId, extractionId);
         } catch (DatabaseServiceException e) {
             throw new InternalServerErrorException(e.getMessage(), e);
         }
+    }
+
+    private static void checkAccessForWorkflowId(int id, Role role) throws DatabaseServiceException {
+        LimsSearchResult result = LIMSInitializationListener.getLimsConnection().getMatchingDocumentsFromLims(
+                Query.Factory.createFieldQuery(LIMSConnection.WORKFLOW_ID_FIELD, Condition.EQUAL, new Object[]{id},
+                        BiocodeService.getSearchDownloadOptions(false, true, false, false)), null, null
+        );
+        List<WorkflowDocument> workflowList = result.getWorkflows();
+        if(workflowList.size() < 1) {
+            throw new NotFoundException("Could not find workflow for id = " + id);
+        }
+        AccessUtilities.checkUserHasRoleForExtractionIds(Collections.singletonList(
+                workflowList.get(0).getWorkflow().getExtractionId()
+        ), role);
     }
 }
